+++ Parameters/MathUtil.h
@@ -139,10 +139,10 @@ struct FFT
       { /* swap a[i] and a[j] */
         t = ar[j - 1];
         ar[j - 1] = ar[i - 1];
-        ar[i - 1] = t;
+        ar[i - 1] = (float)t;
         t = ai[j - 1];
         ai[j - 1] = ai[i - 1];
-        ai[i - 1] = t;
+        ai[i - 1] = (float)t;
       }
       k = NV2; /* bit-reversed counter */
       while (k < j)
@@ -159,8 +159,8 @@ struct FFT
       LE *= 2;  // (LE = 2^L)
       Ur = 1.0f;
       Ui = 0.0f;
-      Wr = cos(M_PI / (float)LE1);
-      Wi = -sin(M_PI / (float)LE1); // Cooley, Lewis, and Welch have "+" here
+      Wr = std::cos((float)M_PI / (float)LE1);
+      Wi = -std::sin((float)M_PI / (float)LE1); // Cooley, Lewis, and Welch have "+" here
       for (j = 1; j <= LE1; j++)
       {
         for (i = j; i <= N; i += LE)
